requirements.txt
asgiref==3.7.2
Django==5.0
mysqlclient==2.2.1
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.9.0

djangorestframework~=3.14.0